src/disease_model/requirements.txt
gitpython~=3.1.2
scipy>=1.4.1
scikit-learn==0.22.1
jupyter
numpy
pandas

